requirements.txt
SQLAlchemy~=2.0.45
pydantic~=2.12.5
passlib~=1.7.4
python-jose~=3.5.0
fastapi~=0.127.0
pytest~=9.0.2
httpx~=0.28.1